% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{mutmat}
\alias{mutmat}
\title{Example binary mutation matrix}
\format{
A matrix with genes as rownames and patient IDs as colnames. Entries are 0/1.
}
\usage{
data(mutmat)
}
\description{
Gene (rows) × patient (columns) 0/1 mutation matrix for examples/tests.
}
\examples{
data(mutmat)
dim(mutmat)
}
\keyword{datasets}
