% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{refgenes}
\alias{refgenes}
\title{Reference gene set used in tests}
\format{
A character vector of gene symbols.
}
\usage{
data(refgenes)
}
\description{
Character vector of gene symbols used as a reference set.
}
\examples{
data(refgenes)
length(refgenes)
}
\keyword{datasets}
